refactor: separate execution history application ownership
